pyproject.toml
[project]
name = "project-coursework-4"
version = "0.1.0"
description = "Coursework Django project"
authors = [
    {name = "Den-is-Ka",email = "D_K81@mail.ru"}
]
requires-python = ">=3.13"
dependencies = [
    "Django (>=5.2,<6.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg[binary] (>=3.2.13,<4.0.0)",
    "django-redis (>=6.0.0,<7.0.0)"
]

# Отключаем упаковку как библиотеку
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "env",
    ".venv",
    ".poetry",
]
ignore = [
    "E501",  # игнорировать длину строки
]
